requirements.txt
# FastAPI and Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Audio Processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.3
ffmpeg-python==0.2.0

# TTS (optional - for local processing)
# torch>=2.0.0
# transformers>=4.30.0
# qwen-tts>=0.1.0

# HTTP Client
httpx==0.26.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
psutil==5.9.8
PyJWT==2.8.0

# Document Parsing
PyPDF2==3.0.1
ebooklib==0.20

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
